Refactored last overlooked test
whcih is moreover confirmiing the theory...
